Separate date implementation into component, need to fix tests for props
